feat: authenticate API calls with either oAuth or API key
Reference: https://www.npmjs.com/package/bcrypt\#why-is-async-mode-recommended-over-sync-mode

+++ src/app/api/v1/organization-members/route.ts
@@ -2,31 +2,67 @@ import { NextResponse } from "next/server";
 import { NextRequest } from "next/server";
 import prisma from "@/lib/prisma";
 import { Prisma } from "@prisma/client";
+import { auth } from "@/auth";
+import checkAuthenticationForAPI from "@/lib/checkAuthenticationForAPI";
 
-export async function GET() {
+export async function GET(req: NextRequest) {
   try {
-    const organizationMembers = await prisma.organization_members.findMany();
+    // Check authentication.
+    const oAuthSession = await auth();
+    const clientSideAuthForAPI = req.headers.get("authorization");
+    const userAuthenticatedID = await checkAuthenticationForAPI(oAuthSession, clientSideAuthForAPI);
+    if (!userAuthenticatedID || userAuthenticatedID === '') {
+      return NextResponse.json(
+        {
+          message: "Authentication failed.",
+        },
+        { status: 401, statusText: "Unauthorized" }
+      );
+    }
+
+    // Proceed if authenticated.
+    const organizationMembers = await prisma.organization_members.findMany({
+      where: {
+        caa_user_id: String(userAuthenticatedID),
+      }
+    });
     return NextResponse.json(organizationMembers, {
       status: 200,
       statusText: "OK",
     });
-  } catch (err) {
-    console.log(err);
-    return NextResponse.json("Error occurred.", {
-      status: 500,
-      statusText: "Internal Server Error",
-    });
+    } catch (err) {
+      console.log(err);
+      return NextResponse.json("Error occurred.", {
+        status: 500,
+        statusText: "Internal Server Error",
+      });
   }
 }
 
 export async function POST(req: NextRequest) {
   const array_organizationMember: Array<Prisma.organization_membersCreateInput> =
     [];
+  let userAuthenticatedID = '';
   try {
+    // Check authentication.
+    const oAuthSession = await auth();
+    const clientSideAuthForAPI = req.headers.get("authorization");
+    userAuthenticatedID = await checkAuthenticationForAPI(oAuthSession, clientSideAuthForAPI);
+    if (!userAuthenticatedID) {
+      return NextResponse.json(
+        {
+          message: "Authentication failed.",
+        },
+        { status: 401, statusText: "Unauthorized" }
+      );
+    }
+ 
+    // Proceed if authenticated.
     const req_payload = await req.json();
     for (const json_req of req_payload) {
       const user: Prisma.organization_membersCreateInput = {
         organizations: json_req["org_id_fk"],
+        caa_user_id: userAuthenticatedID,
         users: json_req["user_email_fk"],
         org_member_role: json_req["org_member_role"],
       };
@@ -54,6 +90,7 @@ export async function POST(req: NextRequest) {
         },
         create: {
           org_id_fk: String(organizationMember["organizations"]),
+          caa_user_id: userAuthenticatedID,
           user_email_fk: String(organizationMember["users"]),
           org_member_role: organizationMember["org_member_role"],
         },
@@ -79,13 +116,40 @@ export async function POST(req: NextRequest) {
 
 export async function DELETE(req: NextRequest) {
   const params = req.nextUrl.searchParams;
+  let userAuthenticatedID = '';
   try {
+    // Check authentication.
+    const oAuthSession = await auth();
+    const clientSideAuthForAPI = req.headers.get("authorization");
+    userAuthenticatedID = await checkAuthenticationForAPI(oAuthSession, clientSideAuthForAPI);
+    if (!userAuthenticatedID) {
+      return NextResponse.json(
+        {
+          message: "Authentication failed.",
+        },
+        { status: 401, statusText: "Unauthorized" }
+      );
+    }
+ 
+    // Proceed if authenticated.
     const delete_all = params.get("delete_all")?.toLowerCase();
     if (delete_all === "true") {
-      const deleteTables = prisma.tables.deleteMany();
-      const deleteProjects = prisma.projects.deleteMany();
+      const deleteTables = prisma.tables.deleteMany({
+        where: {
+          caa_user_id: userAuthenticatedID,
+        }
+      });
+      const deleteProjects = prisma.projects.deleteMany({
+        where: {
+          caa_user_id: userAuthenticatedID,
+        }
+      });
       const deleteOrganizationMembers =
-        prisma.organization_members.deleteMany();
+        prisma.organization_members.deleteMany({
+        where: {
+          caa_user_id: userAuthenticatedID,
+        }
+      });
       const transaction = await prisma.$transaction([
         deleteTables,
         deleteProjects,
@@ -123,6 +187,7 @@ export async function DELETE(req: NextRequest) {
               org_id_fk: org_id_fk,
               user_email_fk: user_email_fk,
             },
+            caa_user_id: userAuthenticatedID,
           },
         }
       );
