+++ src/app/api/v1/tables/route.ts
@@ -2,10 +2,30 @@ import { NextResponse } from "next/server";
 import { NextRequest } from "next/server";
 import prisma from "@/lib/prisma";
 import { Prisma } from "@prisma/client";
+import checkAuthenticationForAPI from "@/lib/checkAuthenticationForAPI";
+import { auth } from "@/auth";
 
-export async function GET() {
+export async function GET(req: NextRequest) {
   try {
-    const tables = await prisma.tables.findMany();
+    // Check authentication.
+    const oAuthSession = await auth();
+    const clientSideAuthForAPI = req.headers.get("authorization");
+    const userAuthenticatedID = await checkAuthenticationForAPI(oAuthSession, clientSideAuthForAPI);
+    if (!userAuthenticatedID || userAuthenticatedID === '') {
+      return NextResponse.json(
+        {
+          message: "Authentication failed.",
+        },
+        { status: 401, statusText: "Unauthorized" }
+      );
+    }
+
+    // Proceed if authenticated.
+    const tables = await prisma.tables.findMany({
+      where: {
+        caa_user_id: userAuthenticatedID,
+      }
+    });
     return NextResponse.json(tables, { status: 200, statusText: "OK" });
   } catch (err) {
     console.log(err);
@@ -18,11 +38,27 @@ export async function GET() {
 
 export async function POST(req: NextRequest) {
   const array_table: Array<Prisma.tablesCreateInput> = [];
+  let userAuthenticatedID = '';
   try {
+    // Check authentication.
+    const oAuthSession = await auth();
+    const clientSideAuthForAPI = req.headers.get("authorization");
+    userAuthenticatedID = await checkAuthenticationForAPI(oAuthSession, clientSideAuthForAPI);
+    if (!userAuthenticatedID) {
+      return NextResponse.json(
+        {
+          message: "Authentication failed.",
+        },
+        { status: 401, statusText: "Unauthorized" }
+      );
+    }
+ 
+    // Proceed if authenticated.
     const req_payload = await req.json();
     for (const json_req of req_payload) {
       const table: Prisma.tablesCreateInput = {
         projects: json_req["project_id_fk"],
+        caa_user_id: userAuthenticatedID,
         table_name: json_req["table_name"],
         table_is_rls_enabled: json_req["table_is_rls_enabled"],
         table_last_updated_on_caa: json_req["table_last_updated_on_caa"],
@@ -52,6 +88,7 @@ export async function POST(req: NextRequest) {
         },
         create: {
           project_id_fk: String(table["projects"]),
+          caa_user_id: userAuthenticatedID,
           table_name: table["table_name"],
           table_is_rls_enabled: table["table_is_rls_enabled"],
           table_last_updated_on_caa: table["table_last_updated_on_caa"],
@@ -78,10 +115,29 @@ export async function POST(req: NextRequest) {
 
 export async function DELETE(req: NextRequest) {
   const params = req.nextUrl.searchParams;
+  let userAuthenticatedID = '';
   try {
+    // Check authentication.
+    const oAuthSession = await auth();
+    const clientSideAuthForAPI = req.headers.get("authorization");
+    userAuthenticatedID = await checkAuthenticationForAPI(oAuthSession, clientSideAuthForAPI);
+    if (!userAuthenticatedID) {
+      return NextResponse.json(
+        {
+          message: "Authentication failed.",
+        },
+        { status: 401, statusText: "Unauthorized" }
+      );
+    }
+ 
+    // Proceed if authenticated.
     const delete_all = params.get("delete_all")?.toLowerCase();
     if (delete_all === "true") {
-      const deleteTables = prisma.tables.deleteMany();
+      const deleteTables = prisma.tables.deleteMany({
+        where: {
+          caa_user_id: userAuthenticatedID,
+        }
+      });
       const transaction = await prisma.$transaction([deleteTables]);
 
       if (process.env.NODE_ENV === "development") {
@@ -114,6 +170,7 @@ export async function DELETE(req: NextRequest) {
             project_id_fk: project_id_fk,
             table_name: table_name,
           },
+          caa_user_id: userAuthenticatedID,
         },
       });
